% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_db_config.R
\name{get_dbInfo}
\alias{get_dbInfo}
\title{get_dbInfo}
\usage{
get_dbInfo(name = NULL)
}
\description{
Get DataBase info from \verb{~/.db.yml}. You need to write config first, see
\code{vignette("database_config")} for details.
}
