% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_mariadb.R, R/open_mysql.R
\name{open_mariadb}
\alias{open_mariadb}
\alias{open_mysql}
\title{Con database connection}
\usage{
open_mariadb(dbname = 1, dbinfo = NULL)

open_mysql(dbname = 1, dbinfo = NULL)
}
\description{
Con database connection
}
\examples{
\dontrun{
dbinfo <- get_dbInfo() # see which db to read
open_mysql()
open_mariadb()
}

}
