% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{make_date}
\alias{make_datetime}
\alias{year}
\alias{month}
\alias{day}
\alias{hour}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lubridate}{\code{\link[lubridate]{day}}, \code{\link[lubridate]{hour}}, \code{\link[lubridate:make_datetime]{make_date}}, \code{\link[lubridate]{make_datetime}}, \code{\link[lubridate]{month}}, \code{\link[lubridate]{year}}}
}}

