% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_time.R
\name{timeinfo}
\alias{timeinfo}
\alias{time2num}
\alias{num2time}
\alias{time2str}
\alias{time_round}
\title{timeinfo}
\usage{
timeinfo(time = Sys.time())

time2num(time, ...)

num2time(num, ...)

time2str(time)

time_round(time = Sys.time())
}
\arguments{
\item{time}{POSIXct or time character, e.g. "1998-01-01 01:00:00"}

\item{num}{difference of \code{time} to \code{1970-01-01} (in second), e.g. \code{1634365535.959902}}
}
\value{
A data.table with the columns of:
\itemize{
\item \code{year}    : int
\item \code{month}   : int
\item \code{day}     : int
\item \code{hour}    : int
\item \code{time}    : time character, e.g. "2021-10-16 14:25:35"
\item \code{timenum} : difference of \code{time} to \code{1970-01-01} (in second), e.g. \code{1634365535.959902}
}
}
\description{
timeinfo
}
\examples{
timeinfo()
time_round()
}
