% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_dbase.R
\name{dbase}
\alias{dbase}
\title{dbase}
\description{
dbase

dbase
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-duckdb_base-new}{\code{dbase$new()}}
\item \href{#method-duckdb_base-open_table}{\code{dbase$open_table()}}
\item \href{#method-duckdb_base-print}{\code{dbase$print()}}
\item \href{#method-duckdb_base-finalize}{\code{dbase$finalize()}}
\item \href{#method-duckdb_base-close}{\code{dbase$close()}}
\item \href{#method-duckdb_base-read_tables}{\code{dbase$read_tables()}}
\item \href{#method-duckdb_base-read_table}{\code{dbase$read_table()}}
\item \href{#method-duckdb_base-read_data}{\code{dbase$read_data()}}
\item \href{#method-duckdb_base-clone}{\code{dbase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_base-new"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_base-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbase$new(db = NULL, table = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_base-open_table"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_base-open_table}{}}}
\subsection{Method \code{open_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbase$open_table(table)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_base-print"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_base-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbase$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_base-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_base-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbase$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_base-close"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_base-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbase$close(force = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_base-read_tables"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_base-read_tables}{}}}
\subsection{Method \code{read_tables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbase$read_tables(tables = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_base-read_table"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_base-read_table}{}}}
\subsection{Method \code{read_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbase$read_table(table = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_base-read_data"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_base-read_data}{}}}
\subsection{Method \code{read_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbase$read_data(..., verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duckdb_base-clone"></a>}}
\if{latex}{\out{\hypertarget{method-duckdb_base-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
