% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_dbase.R
\name{open_mysql}
\alias{open_mysql}
\alias{open_mariadb}
\title{Con database connection}
\usage{
open_mysql(dbname = 1, dbinfo = NULL)

open_mariadb(dbname = 1, dbinfo = NULL)
}
\description{
Con database connection
}
\examples{
\dontrun{
dbinfo = get_dbInfo() # see which db to read
open_mysql()
open_mariadb()
}

}
